# Conflicts:
#	src/main/resources/application.yml

+++ src/main/java/com/kkosoonnae/jpa/projection/StoreListViewProjection.java
@@ -8,7 +8,8 @@ public record StoreListViewProjection(
         Integer storeNo,
         String storeName,
         String img,
-        Integer scope
+        Double scope
+
 
         // ... (목록 조회 때 보여 줄 필드만 작성)
 ) {
